chore: organise components
